Handle atomic copy of download across filesystems
On Linux systems with a separate tmpfs partition at `/tmp`, calling
`os.Rename()` against the created `ioutil.TempFile()` will fail to write
file headers, as the `/home` directory is on a separate partition as the
`/tmp` directory.

This maintains an atomic copy of the temporary download file, and works
across different partitions.